Extract AwsCompactionJobDispatcher
