+++ java/compaction/compaction-job-creation/pom.xml
@@ -52,6 +52,11 @@
             <artifactId>compaction-core</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
+        <dependency>
+            <groupId>sleeper</groupId>
+            <artifactId>compaction-tracker</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
         <!-- Test dependencies -->
         <dependency>
             <groupId>sleeper</groupId>
